@@ -1,3 +1,3 @@
 // This fixed an error related to the CSS and loading gif breaking my Jest test
 // See https://facebook.github.io/jest/docs/en/webpack.html#handling-static-assets
-module.exports = 'test-file-stub';
+export default 'test-file-stub';
